Conflicts:
	persistence/src/main/java/mobi/nowtechnologies/server/persistence/dao/ChartDetailDao.java
	persistence/src/main/java/mobi/nowtechnologies/server/persistence/domain/AccountLog.java
	persistence/src/main/java/mobi/nowtechnologies/server/persistence/domain/User.java
	persistence/src/main/java/mobi/nowtechnologies/server/persistence/domain/UserDeviceDetails.java
	persistence/src/main/java/mobi/nowtechnologies/server/persistence/domain/payment/PSMSPaymentDetails.java
	persistence/src/main/java/mobi/nowtechnologies/server/persistence/domain/payment/PaymentDetails.java
	persistence/src/main/java/mobi/nowtechnologies/server/persistence/repository/UserGroupRepository.java
	persistence/src/main/java/mobi/nowtechnologies/server/persistence/repository/UserRepository.java
	persistence/src/main/resources/META-INF/jpql/chartDetail.xml
	persistence/src/main/resources/sql/release5.0-SNAPSHOT.sql
	persistence/src/test/java/mobi/nowtechnologies/server/persistence/domain/UserTest.java
	persistence/src/test/java/mobi/nowtechnologies/server/persistence/repository/UserRepositoryIT.java
	service/src/main/java/mobi/nowtechnologies/server/service/UserService.java
	service/src/main/java/mobi/nowtechnologies/server/service/impl/UserNotificationServiceImpl.java
	service/src/test/java/mobi/nowtechnologies/server/service/PaymentDetailsServiceTest.java
	service/src/test/java/mobi/nowtechnologies/server/service/UserServiceTest.java
	service/src/test/java/mobi/nowtechnologies/server/service/impl/UserNotificationServiceImplTest.java
	transport/src/main/java/mobi/nowtechnologies/server/job/ForkO2UsersForUpdateJob.java
	transport/src/main/webapp/WEB-INF/conf/log4j.xml

+++ dtos/src/main/java/mobi/nowtechnologies/server/shared/dto/PaymentPolicyDto.java
@@ -18,7 +18,7 @@ public class PaymentPolicyDto {
 
     public static final String PAYMENT_POLICY_DTO = "paymentPolicy";
 
-    private short id;
+    private Integer id;
     private BigDecimal subcost;
     private Integer subweeks;
     private Integer operator;
@@ -67,10 +67,10 @@ public class PaymentPolicyDto {
 		return subweeks == null || subweeks.intValue() == 0;
 	}
     
-    public short getId() {
+    public Integer getId() {
         return id;
     }
-    public void setId(short id) {
+    public void setId(Integer id) {
         this.id = id;
     }
     public BigDecimal getSubcost() {
